[nrf noup] drivers: ieee802154: drop ACK with no timestamp
When a code component, such as printk(), locks interrupts
for the period of a few milliseconds, this may disrupt the
ACK timestamping mechanism implemented in nRF5 802.15.4
radio driver. As a result, an ACK with no timestamp may be
returned to the Zephyr radio driver.

This leads to an assertion failure during timestamp
conversion.

Work around this issue by dropping an ACK with no timestamp
to avoid the device crash or other misbehavior.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>