[nrf fromlist] wifi: shell: Support WPA auto personal security mode
Applications need to run a scan to identify the security
mode before attempting the connection which adds to the time
taken for connection to be established.
To avoid the initial scan, support auto security mode which will
enable STA to choose between WPA, WPA2 and WPA3, based on the
network configuration.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/71044

Signed-off-by: Ravi Dondaputi <ravi.dondaputi@nordicsemi.no>