[nrf noup] boards: Fix MPC configuration for nRF54L15
fixup! [nrf noup] boards: add nrf54l15/nrf54l15/cpuapp/ns as board variant

The nRF54L15 is using the SAU and the MPC to configure the security
attributes for TrustZone. The SPU is only used for the configuration
of the peripherals. This sets the Kconfig configurations for TrustZone
which use the correct region size(based on the MPC) for nRF54L15.

Ref: NCSDK-25023

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>