[nrf noup] soc: nordic: nrf54l: Port Soc.c to be compatible with TFM
In this patch we do several things to be able to configure power and
clocks when NRF_OSCILLATORS and NRF_REGULATORS is secure.

We port soc.c to TF-M so that the TF-M image can do the same
configuration that would normally be done by the Zephyr image.

This is done by explicitly including autoconf.h for instance.

We also ifdef-out the configuration when nordicsemi_nrf54l_init is
invoked from a non-secure image, because the peripherals are secure.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>
(cherry picked from commit e3e053f4b6b58c0dad2a8ad776d1762bb630c621)