[nrf fromlist] bluetooth: tester: Adjust Broadcaster ext adv pars
This change removes the BT_LE_PER_ADV_OPT_USE_TX_POWER option for
extended advertisement in BLE Audio PTS tests.

The Broadcaster setup options was hardcoded to include the
BT_LE_PER_ADV_OPT_USE_TX_POWER option. This causes incompabilities
for any controller not support it, and since the option is not
required for the audio tests it should not be used here.

Upstream PR: zephyrproject-rtos/zephyr#73361

signed-off-by: Frode van der Meeren <frode.vandermeeren@nordicsemi.no>