[nrf fromlist] Bluetooth: Host: Rework enabling of scanner
To make the scanner module more understandable and more streamlined, I
reworked the update mechanism of the scanner. The scanner tracks now the
parameters that were used to enable it and the reason why it is running.
This facilitates state logic and allows other modules to "start the
scanner", altough it is already running.
This is mostly a refactoring and not a functional change.

Added a test to verify the behavior.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/73923

Signed-off-by: Jan Müller <jan.mueller@nordicsemi.no>