[nrf fromlist] nrf_wifi: Remove dependency on OSAL layer handle
Removes the requirement for the different layers in the OS agnostic
parts of the driver having to maintain a handle to the OS interface
layer in order to call the OS interface calls. The OS interface layer
now maitains the handle to OS-specific ops internally and invokes the
appropriate functions.

Fixes SHEL-2639

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/76754

Signed-off-by: Sachin D Kulkarni <sachin.kulkarni@nordicsemi.no>