[nrf fromlist] drivers: serial: nrfx_uarte2: Fix handling of errors
Fix case when interrupt driven API did not enable error interrupt and
error occurs. Previously in that case error state was kept so device
remained in a state in which it was expecting higher layer to read
and clear error state by calling uart_err_check(). If error interrupt
is not enabled it will never occur and in that case error information
shall be dropped.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/74643

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>