[nrf noup] soc: nrf54l: Add custom section for KMU
Add a custom section in the linker which should always
be placed in the top of RAM. This will be used by the
KMU to push keys into it. Since when you provision a key
into the KMU you need to set specific a memory location
for the PUSH operation we need to keep this memory location
static across images/dfus.

The linker script inclusion which places the KMU reserved
buffer on the top of RAM doesn't work for non-XIP builds.
The Zephyr linker script will firstly load the code for
an non-XIP build in RAM and then include this KMU related
linker script which results in an unpredictable placement
of the KMU reserved area and a failed build.

In order to support non-XIP builds the linker file is not
included and the a DTS reserved-memory entry should be used.

To limit the scope, the DTS reserved memory region
is currently only supported for non-XIP builds.

This is a noup since the KMU is not supported upstream.

Ref: NCSDK-25121

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Signed-off-by: Robin Kastberg <robin.kastberg@iar.com>
(cherry picked from commit 673a7d6bb5c611ea09f933a26cf1198de2667a3b)