[nrf fromlist] fs: nvs: Improve nvs_calc_free_space() result precision
The nvs_calc_free_space() function does not return 0 when the NVS
is considered full, because some special ATEs are not taken into account.

This commit takes into account the ATE that is reserved for deletion in
each sector, in addition of the 'GC done' ATE when present.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/74900

Signed-off-by: Adrien Ricciardi <aricciardi@baylibre.com>