[nrf fromtree] fs: nvs: Allow application to switch sector to get free space
Add an API function allowing the application to determine the amount of
free bytes in the current sector.

Add a second API function allowing the application to switch to the next
NVS sector, calling the garbage collector on such sector.

The goal is togive more granularity to the application to control when
the garbage collector is triggered.

(cherry picked from commit b09972b445ef4e85755917ded941e5b1ec4a27d9)

Signed-off-by: Adrien Ricciardi <aricciardi@baylibre.com>