[nrf fromlist] net: ipv6: routing: match destination address to a given interface
Before packet is routed there is a check verifying if the destination
address is not the device's one. However, the check should be limited
to the packet's original interface. Otherwise, packet cannot be routed
if it is destined for an other interface of the device.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/76150

Signed-off-by: Konrad Derda <konrad.derda@nordicsemi.no>