[nrf noup] snippets: Add snippet for the Zephyr Bluetooth Controller
By defining this snippet it becomes simpler to build an application
for the Zephyr Bluetooth Controller as it removes the need
for adding boilerplate overlay files to applications.

This snippet is only needed in NCS because the SoftDevice
Controller is the default and supported controller.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>
(cherry picked from commit e36c39febc1517cbfa9b0dd1132d1fdc0addbbb2)