[nrf noup] soc: nrf54l: Disable CRACEN RAMs on system off
Configure the CRACEN internal RAMs to avoid being retained
during system off. Exiting system off triggers a system level
reset so there is no need for the CRACEN RAMs to be retained
during that.

Also don't retain the CRACEN RAMs in system on idle if CRACEN
is not used at all.

Noup since the upstream code is not TFM aware yet and the
edited function does not exist upstream.

Ref: NCSDK-27427

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>