[nrf fromtree] wifi_mgmt: Add support for configuring PS exit strategy
If AP indicates the presence of buffered traffic, then it is up to the
STA to decide whether to stay in PS or come out of PS, add configuration
options that can be used at runtime to choose this.

This is tagged as "noup" because it's a backport and "fromlist" cannot
be used as it won't apply cleanly.

Signed-off-by: Ajay Parida <ajay.parida@nordicsemi.no>
(cherry picked from commit 0ce5da6da8124dfe760e17e87e60746798aae2a3)
(cherry picked from commit 0dcde521dd07e11b42808125f38f059cb216da86)