[nrf fromlist] modules: hal_nordic: dvfs: queue request when busy
Added one element queue to change frequency setting
function, when previous dvfs sequence is still in
progress. Thanks to this EBUSY status will not be
longer reported and last frequency setting will be
queued and applied as soon as pending change is done.
This is needed by clock control.

Upstream PR #: 78103

Signed-off-by: Lukasz Stepnicki <lukasz.stepnicki@nordicsemi.no>