[nrf fromlist] modules: hal_nordic: dvfs: patch medlow trim entry
DVFS medlow oppoint should use trim entry index 2,
this needs to be updated in hal/hal_nordic.
Currently this is not possible because of time
constraints that is why this temporary patch
is applied. This is the only point where
new_f_trim_entry is used.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/78665

Signed-off-by: Lukasz Stepnicki <lukasz.stepnicki@nordicsemi.no>