drivers/pinctrl.h: Resolve PINCTRL_DT_STATE_INIT Cpp Compatibility
This commit adds C++20 onwards support for the
PINCTRL_DT_STATE_INIT macro. Since C++20, support was
added for designated initializers but with the restriction that
they're ordered. PINCTRL_DT_STATE_INIT initializes a
pinctrl_state struct but the current initializer list is un-ordered,
this PR resolves that for >= C++20 compat

Signed-off-by: Chandler Keep <chandlersamkeep@gmail.com>