[nrf noup] soc: nordic: Disable DCACHE for nrf54h20
When the secure domain services are enabled.

Noup since the symbol SDFW_SERVICES_ENABLED
is not available upstream.

This is required specifically for the PSA crypto
secure service but I cannot use SSF_PSA_CRYPTO_SERVICE_ENABLED
here so I used the SDFW_SERVICES_ENABLED option.

This is temporary since we expect to add support for
data cache later.

This affects the nrf54h20dk cpuapp and cpurad targets.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>