[nrf noup] soc: nordic: Disable DCACHE
When the secure domain services are enabled.

noup since the symbol SDFW_SERVICES_ENABLED
are not available upstream.

This is required specifically for the PSA crypto
secure service but this is the most relevant Kconfig
that I could use.

This is temporary since we expect to add support for
data cache later.

This affects for the following secdom enabled targets:
nrf54h20_cpuapp and cpurad
nrf9280_cpuapp

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>