[nrf noup] boards: Add non-secure target for nrf54L15dk
This adds the nrf54l15dk/nrf54l15/cpuapp/ns board variant to
sdk-zephyr. It allows to build nRF54l15dk with TF-M in nRF Connect SDK.

This is done as a noup as the out-of-tree board extension feature is
not available. When zephyrproject-rtos/zephyr#69548 is completed this
commit can be reverted and the variant be placed in sdk-nrf.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>
(cherry picked from commit 6b4c8404d24374b63546044c2e5752678ede1441)