[nrf fromtree] modules: hostap: Fix memory leak of network
In "connect" all networks are removed and  new network is always added,
but in disconnect the network isn't deleted, so, the memory is
unnecessarily held till next connect. This is not exactly a leak, but if
someone profiles using "kernel heap" then this can be construed as a
leak.

Fix this by removing network during the disconnection (for now "all") so
that the memory can be used by someone else.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>
(cherry picked from commit be65de0d09d19ba2d03e106b81183ec7af996d24)