[nrf fromtree] soc: nordic: nrf53: SOC_NRF53_CPUNET_ENABLE should not depend on !BT
The previous changes in
https://github.com/zephyrproject-rtos/zephyr/pull/74304
assumed that because this is also handled in
`bt_hci_transport_setup` that it shouldn't be done on
initialisation too.

However, if someone wants to develop their own app which
uses BT and also wants to enable the CPUNET by default this
KConfig should be available to them.

Signed-off-by: Sean Madigan <sean.madigan@nordicsemi.no>
(cherry picked from commit e4fa386882269c34cdbaca1ae874de46e2b8aca6)