[nrf noup] drivers: pinctrl_nrf: Add support for EXMIF pins
This is a follow-up to commit 45d827a51a708e54437f5681dc6c4fb77bfb4f52.

Although routing for those pins is configured via UICR, pinctrl still
needs to be involved so that it is possible to set desired drive mode
for them etc.
Add also the missing RWDS pin.

Upstream PR #: 80042
Applied as a [nrf noup] due to a conflict.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>